final1

+++ WWWNhom2Common/src/main/java/com/nhom2/common/entity/CartItem.java
@@ -63,7 +63,7 @@ public class CartItem {
 	}
 
 	@Transactional
-	public float getTotal() {
+	public double getTotal() {
 		return this.product.getDiscountPrice()*quantity;
 	}
